requirements.txt
ultralytics
streamlit
opencv-python-headless
pyngrok
pyyaml
numpy
